Give reconciliation more time
The CRI-O `main` branch and `release-1.29` get synchronized and may use
the same commit. Means if we publish the signed static binary bundle in
the same time via GitHub actions, we risk to upload wrong signatures to
the corresponding artifacts. We avoid this now by giving the jobs a bit
more time in-between.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>